test: msw 핸들러 분리
